+++ tutorials/nservicebus-101/index.md
@@ -3,16 +3,16 @@ title: "NServiceBus 101: Messaging Basics"
 suppressRelated: true
 ---
 
-Welcome to NServiceBus 101! In this self-paced course, you will learn the most important fundamentals of how to get working quickly with NServiceBus.
+In this self-paced course, you will learn the most important fundamentals of how to get working quickly with NServiceBus.
 
 
 ## Objectives
 
-In this course we'll build a simple messaging system from scratch using a retail e-commerce domain.
+In this course we'll build a messaging system using a retail e-commerce domain.
 
 By the end of the course, you will have learned:
 
-* How to set up your development environment for NServiceBus
+* How to set up your development environment
 * What a messaging endpoint is and how to create one
 * How to define messages and message handlers
 * How to send messages between endpoints and control where those messages go
@@ -23,10 +23,10 @@ By the end of the course, you will have learned:
 ## Table of Contents
 
 1. [Lesson 1: Hello world](lesson-1/)
-2. [Lesson 2: Sending a command](lesson-2/)
-3. [Lesson 3: Managing multiple endpoints](lesson-3/)
-4. [Lesson 4: Publishing events](lesson-4/)
-5. [Lesson 5: Retrying errors](lesson-5/)
+1. [Lesson 2: Sending a command](lesson-2/)
+1. [Lesson 3: Managing multiple endpoints](lesson-3/)
+1. [Lesson 4: Publishing events](lesson-4/)
+1. [Lesson 5: Retrying errors](lesson-5/)
 
 
 ## Get Started
