+++ tutorials/nservicebus-101/lesson-2/Snippets/Core_6/Command.cs
@@ -1,0 +1,13 @@
+﻿using NServiceBus;
+
+namespace Core_6
+{
+    #region Command
+
+    public class DoSomething : ICommand
+    {
+        public string SomeProperty { get; set; }
+    }
+
+    #endregion
+}
