+++ tutorials/nservicebus-101/lesson-2/Snippets/Core_6/EmptyHandler.cs
@@ -1,0 +1,17 @@
+﻿using System.Threading.Tasks;
+using NServiceBus;
+
+namespace Core_6.EmptyHandler
+{
+    #region EmptyHandler
+    public class DoSomethingHandler : IHandleMessages<DoSomething>
+    {
+        public Task Handle(DoSomething message, IMessageHandlerContext context)
+        {
+            // Do something with the message here!
+
+            return Task.CompletedTask;
+        }
+    }
+    #endregion
+}
