+++ tutorials/nservicebus-101/lesson-2/Snippets/Core_6/MultiHandler.cs
@@ -1,0 +1,27 @@
+﻿using System;
+using System.Threading.Tasks;
+using NServiceBus;
+
+namespace Core_6.MultiHandler
+{
+    #region MultiHandler
+    public class DoSomethingHandler : 
+        IHandleMessages<DoSomething>,
+        IHandleMessages<DoSomethingElse>
+    {
+        public Task Handle(DoSomething message, IMessageHandlerContext context)
+        {
+            Console.WriteLine("Received DoSomething");
+            return Task.CompletedTask;
+        }
+
+        public Task Handle(DoSomethingElse message, IMessageHandlerContext context)
+        {
+            Console.WriteLine("Received DoSomethingElse");
+            return Task.CompletedTask;
+        }
+    }
+    #endregion
+
+    public class DoSomethingElse : ICommand { }
+}
